backend/requirements-old.txt
audioread>=3.0.1
autopep8>=2.0.1
av>=10.0.0
blessed>=1.20.0
certifi>=2022.12.7
cffi>=1.16.0
charset-normalizer>=3.0.1
click>=8.1.3
coloredlogs>=15.0.1
ctranslate2>=3.22.0
decorator>=5.1.1
faster-whisper>=0.10.0
ffmpeg-python>=0.2.0
filelock>=3.9.0
Flask>=2.2.3
Flask-Cors>=3.0.10
flatbuffers>=23.5.26
fsspec>=2023.10.0
future>=0.18.3
gpustat>=1.0.0
huggingface-hub>=0.19.4
humanfriendly>=10.0
idna>=3.4
importlib-metadata>=6.0.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
joblib>=1.3.2
jsonpickle>=3.0.2
lazy-loader>=0.3
librosa>=0.10.1
llvmlite>=0.41.1
MarkupSafe>=2.1.2
more-itertools>=9.0.0
mpmath>=1.3.0
msgpack>=1.0.7
numba>=0.58.1
numpy>=1.24.2
nvidia-cublas-cu11>=11.10.3.66
nvidia-cuda-nvrtc-cu11>=11.7.99
nvidia-cuda-runtime-cu11>=11.7.99
nvidia-cudnn-cu11>=8.5.0.96
nvidia-ml-py>=11.495.46
onnxruntime>=1.16.3
openai-whisper>=20230124
opus-fast-mosestokenizer>=0.0.8.5
packaging>=23.0
platformdirs>=4.0.0
pooch>=1.8.0
protobuf>=4.25.1
psutil>=5.9.4
pycodestyle>=2.10.0
pycparser>=2.21
PyYAML>=6.0
regex>=2022.10.31
requests>=2.28.2
scikit-learn>=1.3.2
scipy>=1.10.1
six>=1.16.0
soundfile>=0.12.1
soxr>=0.3.7
sympy>=1.12
threadpoolctl>=3.2.0
tokenize-uk>=0.2.0
tokenizers>=0.13.2
tomli>=2.0.1
torch>=1.13.1
tqdm>=4.64.1
transformers>=4.26.1
typing-extensions>=4.5.0
urllib3>=1.26.14
wcwidth>=0.2.6
Werkzeug>=2.2.3
zipp>=3.15.0
